Removed HiddenWordList.h
Removed headerfile and createda text file with a list of words in the new content/wordlists directory.
included Misc FileHelper.h and Misc Paths.h to the cpp
Added member variable Words of type TArray<FString> to the bullcow header
Created FString const in beginplay of the bullcow cpp called WordListPath which stores the path to the newly created txt file called HiddenWordList.txt
Called LoadFileToStringArray from FFileHelper and populated Words array with the text from the txt file.

This will help with compile times and leaves us with a new TArray called Words that contains all the words we will use in the game which is easily edited by simply adding new words or removing old ones from the text file.